// src/api/or-planner/models/MedicalStaff.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OR Planner API
 * Operating Room Planner API
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: filip.chromek@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { MedicalStaffAvailabilityInner } from './MedicalStaffAvailabilityInner';
import {
    MedicalStaffAvailabilityInnerFromJSON,
    MedicalStaffAvailabilityInnerFromJSONTyped,
    MedicalStaffAvailabilityInnerToJSON,
} from './MedicalStaffAvailabilityInner';

/**
 * 
 * @export
 * @interface MedicalStaff
 */
export interface MedicalStaff {
    /**
     * 
     * @type {string}
     * @memberof MedicalStaff
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof MedicalStaff
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof MedicalStaff
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof MedicalStaff
     */
    role: MedicalStaffRoleEnum;
    /**
     * 
     * @type {string}
     * @memberof MedicalStaff
     */
    specialization?: string;
    /**
     * 
     * @type {Array<MedicalStaffAvailabilityInner>}
     * @memberof MedicalStaff
     */
    availability?: Array<MedicalStaffAvailabilityInner>;
    /**
     * 
     * @type {boolean}
     * @memberof MedicalStaff
     */
    archived?: boolean;
}


/**
 * @export
 */
export const MedicalStaffRoleEnum = {
    Doctor: 'doctor',
    Nurse: 'nurse'
} as const;
export type MedicalStaffRoleEnum = typeof MedicalStaffRoleEnum[keyof typeof MedicalStaffRoleEnum];


/**
 * Check if a given object implements the MedicalStaff interface.
 */
export function instanceOfMedicalStaff(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "firstName" in value;
    isInstance = isInstance && "lastName" in value;
    isInstance = isInstance && "role" in value;

    return isInstance;
}

export function MedicalStaffFromJSON(json: any): MedicalStaff {
    return MedicalStaffFromJSONTyped(json, false);
}

export function MedicalStaffFromJSONTyped(json: any, ignoreDiscriminator: boolean): MedicalStaff {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'firstName': json['firstName'],
        'lastName': json['lastName'],
        'role': json['role'],
        'specialization': !exists(json, 'specialization') ? undefined : json['specialization'],
        'availability': !exists(json, 'availability') ? undefined : ((json['availability'] as Array<any>).map(MedicalStaffAvailabilityInnerFromJSON)),
        'archived': !exists(json, 'archived') ? undefined : json['archived'],
    };
}

export function MedicalStaffToJSON(value?: MedicalStaff | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'firstName': value.firstName,
        'lastName': value.lastName,
        'role': value.role,
        'specialization': value.specialization,
        'availability': value.availability === undefined ? undefined : ((value.availability as Array<any>).map(MedicalStaffAvailabilityInnerToJSON)),
        'archived': value.archived,
    };
}